Refactor training process in PINN models
- Removed the adaptive weight update for regularization in the `train_dual_network` function.
- Changed the optimizer from Adam to AdamW to include weight decay.
- Adjusted the update frequency for the optimizer.
- Modified loss computation functions to accept a regularization parameter.
- Updated the logic for updating loss weights based on current losses.
- Changed data types from float64 to float32 for efficiency.
- Deleted obsolete history files and updated binary files in the results directory.

+++ utils/infinite.py
@@ -60,7 +60,7 @@ def generate_dataset_inf(
     n_pde=10_000,
     n_grid=300,
     device="cpu",
-    dtype=torch.float64,
+    dtype=torch.float32,
     plot=False,           
 ):
     xmin, xmax = domain
@@ -247,7 +247,7 @@ def evaluate_model_inf(
 
     X_test = torch.tensor(
         np.column_stack((X.ravel(), Y.ravel())),
-        dtype=torch.float64,
+        dtype=torch.float32,
         device=device,
     )
